refactor:abstract components and  solve comments
