+++ __tests__/components/AdminNavbarDesktop.test.tsx
@@ -1,5 +1,5 @@
 import { render } from '../../utils/render-test';
-import AdminNavbarDesktop from '@/components/AdminNavbarDesktop';
+import AdminNavbarDesktop from '@/pages/admin/AdminNavbarDesktop';
 import { navItems } from '@/utils/navItems';
 
 describe('AdminNavbarDesktop Snapshot', () => {
