+++ __tests__/components/AdminTable.test.tsx
@@ -1,4 +1,4 @@
-import { AdminTable } from '@/components/AdminTable';
+import { AdminTable } from '@/pages/admin/AdminTable';
 import { render } from '@/utils/render-test';
 
 // Mock data for testing
