+++ components/Layout.tsx
@@ -1,7 +1,5 @@
-// components/Layout.tsx
+import AdminNavbar from '@/pages/admin/AdminNavbar';
 import React from 'react';
-import AdminNavbar from './AdminNavbar';
-
 export default function Layout({ children }: { children: React.ReactNode }) {
     return (
         <div>
