+++ pages/admin/constraints/adminMessages.ts
@@ -1,0 +1,8 @@
+export const adminTableMessages = {
+    
+    emptyData: {
+      es: "No hay datos disponibles"
+    }
+  };
+  
+
